Rename auth components module
